Adaptative and usable Menus !

+++ WizWorld/Include/Entity.h
@@ -9,17 +9,17 @@ class Entity
 {
 public:
 	Entity();
-	Entity(float speed, float x_shape, float y_shape);
+	Entity(double speed, double x_shape, double y_shape);
 	virtual ~Entity();
 	/* Move the entity; x and y are in {-1, 0, 1} */
-	virtual void move(const float& dt, const float x_motion, const float y_motion);
+	virtual void move(const double& dt, const double x_motion, const double y_motion);
 
-	virtual void update(const float &dt);
+	virtual void update(const double &dt);
 	virtual void render(sf::RenderTarget* target);
 
 protected:
 	sf::RectangleShape m_shape;
-	float m_speed;
+	double m_speed;
 };
 
 #endif //!_ENTITY_H_
